Entire Solution - Replaced the Logger class with Log4Net which is threadsafe. and added a custom Appender so that we can capture data to the editor as well as logs. Logs will be outputed to log\MudEngine.log which can be opened in any text editor. Currently everythign is set to log all levels but can be changed in the log4net.config  to specify WARN, ERROR etc.
Also added a little bit more logging to the classes, and enabled log4net on script classes as well so you can Log from your script files without an issue.

Engine - Added Crypt class to hash passwords that are set in game, this way no cleartext passwords exist, anything saved to file will be hashed so no one is able to determine the password the user sent in.